refactor(matrices): move fill and zip_row_op to RepMatrix
